Continue refactoring
