+++ dockstore-webservice/src/main/java/io/dockstore/webservice/core/User.java
@@ -39,6 +39,8 @@ import javax.persistence.Table;
 
 import com.fasterxml.jackson.annotation.JsonIgnore;
 import com.fasterxml.jackson.annotation.JsonProperty;
+import io.dockstore.webservice.helpers.GitHubSourceCodeRepo;
+import io.dockstore.webservice.jdbi.TokenDAO;
 import io.swagger.annotations.ApiModel;
 import io.swagger.annotations.ApiModelProperty;
 
@@ -110,6 +112,18 @@ public class User implements Principal {
         groups = new HashSet<>(0);
         entries = new HashSet<>(0);
         starredEntries = new LinkedHashSet<>();
+    }
+
+    /**
+     * Updates the given user with metadata from Github
+     *
+     * @param tokenDAO
+     * @return updated user
+     */
+    public void updateUserMetadata(final TokenDAO tokenDAO) {
+        Token githubToken = tokenDAO.findGithubByUserId(getId()).get(0);
+        GitHubSourceCodeRepo gitHubSourceCodeRepo = new GitHubSourceCodeRepo(getUsername(), githubToken.getContent(), null);
+        gitHubSourceCodeRepo.getUserMetadata(this);
     }
 
     @JsonProperty
