+++ dockstore-webservice/src/main/java/io/dockstore/webservice/core/Workflow.java
@@ -16,9 +16,9 @@
 
 package io.dockstore.webservice.core;
 
-import com.fasterxml.jackson.annotation.JsonProperty;
-import io.swagger.annotations.ApiModel;
-import io.swagger.annotations.ApiModelProperty;
+import java.util.Set;
+import java.util.SortedSet;
+import java.util.TreeSet;
 
 import javax.persistence.Column;
 import javax.persistence.DiscriminatorValue;
@@ -34,9 +34,10 @@ import javax.persistence.OneToMany;
 import javax.persistence.OrderBy;
 import javax.persistence.Table;
 import javax.persistence.UniqueConstraint;
-import java.util.Set;
-import java.util.SortedSet;
-import java.util.TreeSet;
+
+import com.fasterxml.jackson.annotation.JsonProperty;
+import io.swagger.annotations.ApiModel;
+import io.swagger.annotations.ApiModelProperty;
 
 /**
  * This describes one workflow in the dockstore, extending Entry with the fields necessary to describe workflows.
@@ -127,6 +128,30 @@ public class Workflow extends Entry<Workflow, WorkflowVersion> {
         this.setMode(workflow.getMode());
         this.setWorkflowName(workflow.getWorkflowName());
         this.setPath(workflow.getPath());
+    }
+
+    /**
+     * Copies some of the attributes of the source workflow to the target workflow
+     * There are two of these which seems redundant.
+     *
+     * @param targetWorkflow
+     * @deprecated seems to overlap with {@link #update(Workflow)} , it is not clear why both exist
+     */
+    @Deprecated
+    public void copyWorkflow(Workflow targetWorkflow) {
+        targetWorkflow.setPath(getPath());
+        targetWorkflow.setIsPublished(getIsPublished());
+        targetWorkflow.setWorkflowName(getWorkflowName());
+        targetWorkflow.setAuthor(getAuthor());
+        targetWorkflow.setEmail(getEmail());
+        targetWorkflow.setDescription(getDescription());
+        targetWorkflow.setLastModified(getLastModified());
+        targetWorkflow.setOrganization(getOrganization());
+        targetWorkflow.setRepository(getRepository());
+        targetWorkflow.setGitUrl(getGitUrl());
+        targetWorkflow.setDescriptorType(getDescriptorType());
+        targetWorkflow.setDefaultVersion(getDefaultVersion());
+        targetWorkflow.setDefaultTestParameterFilePath(getDefaultTestParameterFilePath());
     }
 
     @JsonProperty
